Add files via upload
The goal of this lab is to explore some of the basic C programming concepts along with Arduino coding targeted towards an LCD display and button input platform that will be useful for all the other labs in this course. Additionally, you will gain experience with pointers, arrays, and linked list data structures taught in MREN 178 lectures.